requirements.txt
streamlit
pandas
plotly
scikit-learn
joblib
SQLAlchemy
psycopg2-binary
pandasai
python-dotenv
matplotlib
sqlglot
google-generativeai